add new length regulator
