+++ nemo/collections/tts/modules/fastspeech2s.py
@@ -110,34 +110,6 @@ class FFTBlocks(nn.Module):
         return output, non_pad_mask
 
 
-class LengthRegulator(nn.Module):
-    def forward(self, x, dur):
-        output = []
-        for x_i, dur_i in zip(x, dur):
-            expanded = self.expand(x_i, dur_i)
-            output.append(expanded)
-        output = self.pad(output)
-        return output
-
-    def expand(self, x, dur):
-        output = []
-        for i, frame in enumerate(x):
-            expanded_len = int(dur[i] + 0.5)
-            expanded = frame.expand(expanded_len, -1)
-            output.append(expanded)
-        output = torch.cat(output, 0)
-        return output
-
-    def pad(self, x):
-        output = []
-        max_len = max([x[i].size(0) for i in range(len(x))])
-        for i, seq in enumerate(x):
-            padded = F.pad(seq, [0, 0, 0, max_len - seq.size(0)], 'constant', 0.0)
-            output.append(padded)
-        output = torch.stack(output)
-        return output
-
-
 class FFTBlock(torch.nn.Module):
     """FFT Block"""
